requirements.txt
py-clob-client>=0.0.1
python-dotenv>=1.0.0
web3>=6.0.0
requests>=2.31.0
flask>=3.0.0
